Fix scheduling of message in PredisSetDriver
The parameters of zadd were not in sync with the Predis library.
Second parameter should be $membersAndScoresDictionary which is
an array with the key-value pairs containing message-score.

It is also allowed to use two parameters, however the furst one
is still expected to be a message.

https://github.com/predis/predis/blob/4c1aada5aebf0afde266b512ed069f822148c3a2/src/Command/Redis/ZADD.php

The original implementation passed a score as a first one, which
caused that server tried to use contents of the message as float-based
score. This resulted in "not a valid float" error.

tomaj/hermes#48